+++ jvm-packages/xgboost4j/src/main/java/ml/dmlc/xgboost4j/java/XGBoostJNI.java
@@ -71,7 +71,7 @@ class XGBoostJNI {
   public final static native int XGDMatrixCreateFromMat(float[] data, int nrow, int ncol,
                                                         float missing, long[] out);
 
-  public final static native int XGDMatrixCreateFrom2DMat(float[][] data, int nrow, int ncol,
+  public final static native int XGDMatrixCreateFrom2DMat(float[][] data, long nrow, int ncol,
                                                         float missing, long[] out);
 
   public final static native int XGDMatrixSliceDMatrix(long handle, int[] idxset, long[] out);
